4-MV-Considerations/requirements.txt
kafka-python
clickhouse-connect
tabulate
